refactor: replace Navigation component with specific AuthNavigation and LandingNavigation in Home and Landing pages
